add groups

+++ src/main/resources/db/users.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    id INT PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     username VARCHAR,
     first_name VARCHAR,
     last_name VARCHAR,
